Protected course editing & removing

+++ routes/courses.js
@@ -3,41 +3,67 @@ const Course = require("../models/course");
 const auth = require("../middleware/auth");
 const router = Router();
 
-router.get("/", async (req, res) => {
-  const courses = await Course.find()
-    .populate("userId", "email name")
-    .select("price title img");
+function isOwner(course, req) {
+  return course.userId.toString() === req.user._id.toString();
+}
 
-  res.render("courses", {
-    title: "Courses",
-    isCourses: true,
-    courses,
-  });
+router.get("/", async (req, res) => {
+  try {
+    const courses = await Course.find()
+      .populate("userId", "email name")
+      .select("price title img");
+
+    res.render("courses", {
+      title: "Courses",
+      isCourses: true,
+      userId: req.user ? req.user._id.toString() : null,
+      courses,
+    });
+  } catch (error) {
+    console.log(error);
+  }
 });
 
 router.get("/:id/edit", auth, async (req, res) => {
-  if (!req.query.allow) {
-    return res.redirect("/");
+  try {
+    if (!req.query.allow) {
+      return res.redirect("/");
+    }
+
+    const course = await Course.findById(req.params.id);
+
+    if (!isOwner(course, req)) {
+      return res.redirect("/courses");
+    }
+
+    res.render("course-edit", {
+      title: `Edit ${course.name}`,
+      course,
+    });
+  } catch (error) {
+    console.log(error);
   }
-
-  const course = await Course.findById(req.params.id);
-
-  res.render("course-edit", {
-    title: `Edit ${course.name}`,
-    course,
-  });
 });
 
 router.post("/edit", auth, async (req, res) => {
-  const { id } = req.body;
-  delete req.body.id;
-  await Course.findByIdAndUpdate(id, req.body);
-  res.redirect("/courses");
+  try {
+    const { id } = req.body;
+    delete req.body.id;
+    const course = await Course.findById(id);
+    if (!isOwner(course, req)) {
+      return res.redirect("/courses");
+    }
+    Object.assign(course, req.body);
+    // await Course.findByIdAndUpdate(id, req.body);
+    res.redirect("/courses");
+  } catch (error) {
+    console.log(error);
+  }
 });
 
 router.post("/remove", auth, async (req, res) => {
   try {
-    await Course.deleteOne({ _id: req.body.id });
+    await Course.deleteOne({ _id: req.body.id, userId: req.user._id });
     res.redirect("/courses");
   } catch (error) {
     console.log(error);
@@ -45,12 +71,16 @@ router.post("/remove", auth, async (req, res) => {
 });
 
 router.get("/:id", async (req, res) => {
-  const course = await Course.findById(req.params.id);
-  res.render("course", {
-    layout: "empty",
-    title: `Course ${course.name}`,
-    course,
-  });
+  try {
+    const course = await Course.findById(req.params.id);
+    res.render("course", {
+      layout: "empty",
+      title: `Course ${course.name}`,
+      course,
+    });
+  } catch (error) {
+    console.log(error);
+  }
 });
 
 module.exports = router;
